feat(api): add chat routes and consolidate shared infra into @notra/ai
Adds the public Chats API to apps/api (Hono OpenAPI):
- GET  /v1/chats                 list chats
- GET  /v1/chats/{chatId}        get one chat with messages
- POST /v1/chats                 start a new chat, stream the reply
- POST /v1/chats/{chatId}        post a message to an existing chat, stream the reply

Streaming uses the existing standalone-chat orchestration (direct or qstash workflow,
matching the dashboard). The minted chat id is returned via the X-Chat-Id response
header and embedded in the start chunk's messageMetadata.chatId.

Also consolidates shared infrastructure that was siloed in apps/dashboard into the
@notra/ai package so both apps share one source of truth:
- @notra/ai/billing/{autumn,token-pricing,features}
- @notra/ai/qstash/{client,triggers}
- @notra/ai/realtime, @notra/ai/evlog
- @notra/ai/chat/{history (createChatSession, getChatSession), integrations-cache}
- @notra/ai/integrations/{github,linear}
- @notra/ai/crypto/token-encryption, @notra/ai/utils/{redis,url}

Dashboard call sites kept working via thin re-export shims at the original paths.

OpenAPI spec exposes all four chat operations; apps/docs/docs.json sidebar adds a
"Chats" group under API Reference > Endpoints.

Type checks: apps/api, packages/ai, apps/dashboard all clean.
