+++ apps/api/src/routes/chat-workflow.ts
@@ -1,0 +1,9 @@
+import { OpenAPIHono } from "@hono/zod-openapi";
+import type { Context } from "hono";
+import { chatWorkflowHandler } from "../lib/chat/workflow-handler";
+
+export const chatWorkflowRoutes = new OpenAPIHono();
+
+chatWorkflowRoutes.post("/workflows/chat", (c: Context) =>
+  chatWorkflowHandler(c)
+);
