+++ apps/dashboard/src/lib/billing/autumn.ts
@@ -1,7 +1,1 @@
-import { Autumn } from "autumn-js";
-
-const AUTUMN_SECRET_KEY = process.env.AUTUMN_SECRET_KEY;
-
-export const autumn = AUTUMN_SECRET_KEY
-  ? new Autumn({ secretKey: AUTUMN_SECRET_KEY })
-  : null;
+export { autumn } from "@notra/ai/billing/autumn";
