+++ apps/dashboard/src/lib/crypto/token-encryption.ts
@@ -1,17 +1,1 @@
-import {
-  decodeIntegrationEncryptionKey,
-  decryptIntegrationSecret,
-  encryptIntegrationSecret,
-} from "@notra/db/utils/integration-encryption";
-
-function getEncryptionKey() {
-  return decodeIntegrationEncryptionKey(process.env.INTEGRATION_ENCRYPTION_KEY);
-}
-
-export function encryptToken(token: string) {
-  return encryptIntegrationSecret(token, getEncryptionKey());
-}
-
-export function decryptToken(encryptedToken: string) {
-  return decryptIntegrationSecret(encryptedToken, getEncryptionKey());
-}
+export { decryptToken, encryptToken } from "@notra/ai/crypto/token-encryption";
