+++ packages/ai/package.json
@@ -9,6 +9,8 @@
     "./gateway": "./src/gateway.ts",
     "./model": "./src/model.ts",
     "./observability": "./src/observability.ts",
+    "./evlog": "./src/evlog.ts",
+    "./realtime": "./src/realtime.ts",
     "./agents/*": "./src/agents/*.ts",
     "./orchestration/*": "./src/orchestration/*.ts",
     "./tools/*": "./src/tools/*.ts",
@@ -22,7 +24,11 @@
     "./schemas/*": "./src/schemas/*.ts",
     "./utils/*": "./src/utils/*.ts",
     "./constants/*": "./src/constants/*.ts",
-    "./chat/*": "./src/chat/*.ts"
+    "./chat/*": "./src/chat/*.ts",
+    "./billing/*": "./src/billing/*.ts",
+    "./qstash/*": "./src/qstash/*.ts",
+    "./integrations/*": "./src/integrations/*.ts",
+    "./crypto/*": "./src/crypto/*.ts"
   },
   "dependencies": {
     "@ai-sdk-tools/cache": "^1.2.0",
@@ -31,11 +37,15 @@
     "@notra/db": "workspace:*",
     "@octokit/core": "^7.0.6",
     "@supermemory/tools": "^1.3.64",
+    "@upstash/qstash": "2.9.0-rc",
+    "@upstash/realtime": "^1.0.3",
     "@upstash/redis": "^1.35.8",
+    "@upstash/workflow": "^1.1.0",
     "ai": "^6.0.168",
+    "autumn-js": "^1.2.7",
     "dedent": "^1.7.1",
     "drizzle-orm": "^0.45.2",
-    "evlog": "^2.10.0",
+    "evlog": "^2.13.0",
     "gray-matter": "^4.0.3",
     "marked": "^17.0.1",
     "nanoid": "^5.1.6",
